pyproject.toml
[project]
name = "activetrack"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "garminconnect (>=0.2.30,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask (>=3.0.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
